Fix upper manual key-split -- #63
This is reformatted patch from @johnfolk, who describes it:

The 'problem' was that the MIDI note range always was starting from the
non-transposed left edge (24 or max (left, 36)) to the non-transposed split
point. If there were unused keys to the left (i.e. less than 24 or 36)
they were not used even if they might map to B3 keys and were above 'left'.
The last region had the same issue but the code was somewhat different before
leading to somewhat different effects but the casuse was similar, the unused
note numbers on the left edge.

The strategy now is to use all possible MIDI note numbers by adjusting the
`first_MIDI_Note` and transpose to maintain the difference between them but
move the first MIDI lower as much as possible. Thereby extending the range
of note numbers that would be attmpted to match the B3's range. Then adjust
if needed the start MIDI and key to match the left boundary from the previous
split point (as before). The main thing is to be sure that the mapping constant:

  `offset = trnp + firstkey - first_MIDI_Note`

has not changed and that we respect the left limit on first_MIDI_Note. While
increasing the range to the left to include any unused MIDI note numbers that
would map to valid keys.

This works correctly for a reasonble range of
`m->transpose + m->nshA_PL, m->splitA_PL`. Note numbers will stay >= 0.
If transposes get wild they simply will not map to any keys. At the end of the
day the Midi+transpose must hit the limited key ranges or the table seems to
get set to 255. Which is fine.